chore: Add unit test for date parsing
refactored all dates functions into date-utils
related to #1426
